Changed variable with type long to size_t

+++ cbase64.h
@@ -4,6 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stddef.h>
 
 static char base64_encode_lut[64] = {
 	'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H',
@@ -32,11 +33,11 @@ static int base64_decode_lut[123] = {
 
 static int base64_padding[3] = {0, 2, 1};
 
-char* base64_encode_b64ch(char* bin, long* text_len, long* text_size, char* buffer){
+char* base64_encode_b64ch(char* bin, size_t* text_len, size_t* text_size, char* buffer){
 	int padding = *text_len % 3;
-	long block_size = (*text_size + ((*text_size % 6) > 0 ? 6 - (*text_size % 6) : 0)) / 6;
+	size_t block_size = (*text_size + ((*text_size % 6) > 0 ? 6 - (*text_size % 6) : 0)) / 6;
 	char* b64_text = (char*)malloc((block_size * sizeof(char)) + base64_padding[padding] + 1);
-	for(long i = 0;i < block_size;i++){
+	for(size_t i = 0;i < block_size;i++){
 		memset(buffer, '0', 8);
 		memcpy(&buffer[2], &bin[i*6], (*text_size - (i * 6)) < 6 ? (*text_size - (i * 6)): 6);
 		b64_text[i] = base64_encode_lut[strtol(buffer, NULL, 2)];
@@ -48,10 +49,10 @@ char* base64_encode_b64ch(char* bin, long* text_len, long* text_size, char* buff
 
 char* base64_encode(char* text){
 	char buffer[8];
-	long text_len = strlen(text);
-	long text_size = text_len * 8, n, j;
+	size_t text_len = strlen(text);
+	size_t text_size = text_len * 8, n, j;
 	char *bin = (char*)malloc(text_size);
-	for(long i = 0;i < text_len;i++){
+	for(size_t i = 0;i < text_len;i++){
 		memset(buffer, '0', 8);
 		n = (int)text[i],
 		j = 7;
@@ -68,8 +69,8 @@ char* base64_encode(char* text){
 char* base64_encode_file(FILE *file){
 	char buffer[8];
 	fseek(file, 0, SEEK_END);
-	long text_len = ftell(file);
-	long text_size = text_len * 8;
+	size_t text_len = ftell(file);
+	size_t text_size = text_len * 8;
 	char *bin = (char*)malloc(text_size);
 	fseek(file, 0, SEEK_SET);
 	int ch, x = 0, j;
@@ -90,12 +91,12 @@ char* base64_encode_file(FILE *file){
 
 char* base64_decode_b64ch(char* b64_text){
 	char buffer[8];
-	long text_len = strlen(b64_text);
+	size_t text_len = strlen(b64_text);
 	check_padding:
 		if(b64_text[text_len - 1] == '='){text_len--; goto check_padding;}
 	int n, j;
 	char* bin = (char*)malloc(text_len * 6);
-	for(long i = 0;i < text_len;i++){
+	for(size_t i = 0;i < text_len;i++){
 		memset(buffer, '0', 8);
 		n = base64_decode_lut[(int)b64_text[i]],
 		j = 7;
@@ -112,9 +113,9 @@ char* base64_decode_b64ch(char* b64_text){
 char* base64_decode(char* b64_text){
 	char buffer[8];
 	char *bin = base64_decode_b64ch(b64_text);
-	long data_len = strlen(bin) >> 3;
+	size_t data_len = strlen(bin) >> 3;
 	char *text = (char*)malloc(data_len);
-	for(long i = 0;i < data_len;i++){
+	for(size_t i = 0;i < data_len;i++){
 		memcpy(buffer, &bin[i*8], 8);
 		text[i] = (char)strtol(buffer, NULL, 2);
 	}
@@ -125,8 +126,8 @@ char* base64_decode(char* b64_text){
 int base64_decode_file(char* b64_text, FILE* file){
 	char buffer[8];
 	char *bin = base64_decode_b64ch(b64_text);
-	long data_len = strlen(bin) >> 3;
-	for(long i = 0;i < data_len;i++){
+	size_t data_len = strlen(bin) >> 3;
+	for(size_t i = 0;i < data_len;i++){
 		memcpy(buffer, &bin[i*8], 8);
 		fputc(strtol(buffer, NULL, 2), file);
 	}
